refactor: decompose schema generation pipeline
Split schema generation into typed stage modules for support, analysis,
workflow, and stage models while keeping the public schema_generation
facade stable.
